Convert hosting device's created_at attribute to datetime object only if needed
Initially the hosting device['created_at'] atribute is coming as a string. Since we use it all the time, it is converted to a datetime object. But when the router_deleted action is happening, we inadvertently converts it again causing the TypeError. An instanceof() check is added to
avoid this.

Closes defect: DE1016

Change-Id: Ia5a71c2498cdd8ae1d5b815db6a987575a29384c