Retry if HA group info is not found for HA enabled router
This patch changes how HA information is added to ports for
an HA enabled router. If the HA group for a router and a particular
subnet is not found in the DB, the plugin will retry the DB
query a few times in hope that the HA group DB entry will be
commited to the DB during those retries.

The cfg agent is modified so that it handles cases where the
above retry mechanism in the plugin fails. The plugin is then
not able to set the HA data when sending the router dict to
the cfg agent. In such cases where the HA is enabled and no
HA_INFO is set, the cfg agent raise an error and retries.

The patch also makes a few minor code cleanups of the code
handling the HA group info, in particular make use of a constant
defined in the HA extension instead of its string value directly.

Closes bug: DE936

Change-Id: I3152fe7a2bd5e31db13fa7d3f63485cffd74b0a4