Splitting up atomic calls to APIs.
This splits up the calls to the superclass call
for creating routers. It breaks calls for router
creation with gateways into separate calls to the
superclass, in order to avoid problems where the
create_port_(pre|post)commit calls for the router
gateway port look for the newly created router,
but fail for lack of a host binding.

Signed-off-by: Thomas Bachman <tbachman@yahoo.com>