UCSM MD scale improvement, selectively process port updates.
Currently, the UCSM mechanism driver handles nearly every
update_port_postcommit by, at a minimum, connecting to the UCSM
and checking various conditions via get operations.  The
neutron ML2 manager invokes this driver method very often and
with port contents that are not actionable or redundant given
prior ports updated in the processing flow.

This change is to make the UCSM driver's update_port_postcommit
take the same strategy as the Nexus driver--specifically, only
process active ports that are associated with compute nodes,
router HA ports, and dhcp-ports.  If additional port types are
to be supported in the future, they will need to be explicitly
added to the CiscoUcsmMechanismDriver's
_is_supported_deviceowner method.

Change-Id: I13c575370156808e351ed589d68411a5d2482d2f