Fix subnet selection for global routers
This fixes the selection of the proper subnet when creating
the global and logical global routers. When multiple subnets
exist on the external network, the code would select the subnet
with the lower UUID (i.e. the one appearing first in a list).
This patch fixes it so that the global and logical global
routers use the same subent as the user's router.

Signed-off-by: Thomas Bachman <tbachman@yahoo.com>