Replace ciscoconfparser with HTParser
This commit introduces a simple hierarchical text parser called
HTParser. It uses indent levels to determine parent child hierarchy,
specifically in IOS config strings. In this application we use the
parser to determine the presence of a specific config or sub-configs
in a config resource.

This in-tree module replaces the use of the ciscoconfparse library.
Only a subset of functionality provided by ciscoconfparse is used
currently and having a module in-tree for the same gives us more
flexibility.

This closes #168

Signed-off-by: Thomas Bachman <tbachman@yahoo.com>