Catch StaleDataError in disassociate_floatingips
Sometimes deletion of port during disassociate_floatingips
and update of same port are being executed concurrently.
This leads to a StaleDataError but that error does not
mean that something is actually broken in this case.

This patch therefore catches that exception and just logs
it so the execution can continue without tracebacks.

Closes defect: DE2297

Change-Id: Iae7182ffbeba1f8eaa58404a309818f9b85a817c