Adds hosting device simulation capability to config agent
The config agent is given functionality to simulate all
hosting devices to look like IOS XE devices. This is for
the convenience of developers to allow them to do basic
testing without access to a testbed with real hosting
devices.

The dataplane of the hosting devices is NOT simulated,
except for the health check pings, which are mocked to
look as if they performed. Those mocked pings can be
controlled so that they either appears to succeed or
to fail.

Change-Id: I3825efa5bc93a19c76366b9ef3317ceaa7735852